+++ netZooPy/lioness/lioness.py
@@ -1,5 +1,6 @@
 from __future__ import print_function
 import os, os.path, sys
+#from netZooPy.command_line import lioness
 import numpy as np
 import pandas as pd
 from .timer import Timer
@@ -13,51 +14,87 @@ from joblib import wrap_non_picklable_objects
 from netZooPy.panda.calculations import compute_panda
 
 class Lioness(Panda):
-    """
-    Description:
-        Using LIONESS to infer single-sample gene regulatory networks.
+    """ Using LIONESS to infer single-sample gene regulatory networks.
         1. Reading in PANDA network and preprocessed middle data
         2. Computing coexpression network
         3. Normalizing coexpression network
         4. Running PANDA algorithm
         5. Writing out LIONESS networks
 
-    Inputs:
-        Panda: Panda object.
+    Parameters
+    ----------
 
-    Methods:
-        __init__              : Intialize instance of Lioness class and load data.
-        __lioness_loop        : The LIONESS algorithm.
-        save_lioness_results  : Saves LIONESS network.
+            obj             : object
+                PANDA object, generated with keep_expression_matrix=True.
+            computing       : str
+                'cpu' uses Central Processing Unit (CPU) to run PANDA
+                'gpu' use the Graphical Processing Unit (GPU) to run PANDA
+            precision       : str
+                'double' computes the regulatory network in double precision (15 decimal digits).
+                'single' computes the regulatory network in single precision (7 decimal digits) which is fastaer, requires half the memory but less accurate.
+            start           : int
+                Index of first sample to compute the network.
+            end             : int
+                Index of last sample to compute the network.
+            save_dir        : str
+                Directory to save the networks.
+            save_fmt        : str
+                Save format.
+                - '.npy': (Default) Numpy file.
+                - '.txt': Text file.
+                - '.mat': MATLAB file.
+            output          : str
+                - 'network' returns all networks in a single edge-by-sample matrix (lioness_obj.total_lioness_network is the unlabeled variable and lioness_obj.export_lioness_results is the row-labeled variable). For large sample sizes, this variable requires large RAM memory.
+                - 'gene_targeting' returns gene targeting scores for all networks in a single gene-by-sample matrix (lioness_obj.total_lioness_network).
+                - 'tf_targeting' returns tf targeting scores for all networks in a single gene-by-sample matrix (lioness_obj.total_lioness_network).
+            alpha            : float
+                learning rate, set to 0.1 by default but has to be changed manually to match the learning rate of the PANDA object.
+            save_single: bool
+                when set to True it will save each lioness network with its sample name inside the lioness output folder
+            export_filename: str
+                if passed, the final lioness table will be saved with all tf-gene edges as dataframe index and 
+                samples as column name
+    Returns
+    --------
+    export_lioness_results : _
+        Depeding on the output argument, this can be either all the lioness 
+        networks or their gene/tf targeting scores.
 
-    Example:
-        from netZooPy.lioness.lioness import Lioness
-        To run the Lioness algorithm for single sample networks, first run PANDA using the keep_expression_matrix flag, then use Lioness as follows:
-        panda_obj = Panda('../../tests/ToyData/ToyExpressionData.txt', '../../tests/ToyData/ToyMotifData.txt', '../../tests/ToyData/ToyPPIData.txt', remove_missing=False, keep_expression_matrix=True)
-        lioness_obj = Lioness(panda_obj)
+    Notes
+    -------
+    Example on how to use Lioness and plot the network
 
-        Save Lioness results:
-        lioness_obj.save_lioness_results('Toy_Lioness.txt')
-        Return a network plot for one of the Lioness single sample networks:
-        plot = AnalyzeLioness(lioness_obj)
-        plot.top_network_plot(column= 0, top=100, file='top_100_genes.png')
+        >>> from netZooPy.lioness.lioness import Lioness
+        >>> #To run the Lioness algorithm for single sample networks, first run PANDA using the keep_expression_matrix flag, then use Lioness as follows:
+        >>> panda_obj = Panda('../../tests/ToyData/ToyExpressionData.txt', '../../tests/ToyData/ToyMotifData.txt', '../../tests/ToyData/ToyPPIData.txt', remove_missing=False, keep_expression_matrix=True)
+        >>> lioness_obj = Lioness(panda_obj)
 
-        Example lioness output:
-        Sample1 Sample2 Sample3 Sample4
-        -------------------------------
-        -0.667452814003	-1.70433776179	-0.158129613892	-0.655795512803
-        -0.843366539284	-0.733709815256	-0.84849895139	-0.915217389738
-        3.23445386464	2.68888472802	3.35809757371	3.05297381396
-        2.39500370135	1.84608635425	2.80179804094	2.67540878165
-        -0.117475863987	0.494923925853	0.0518448588965	-0.0584810456421
+        >>> #Save Lioness results:
+        >>> lioness_obj.save_lioness_results('Toy_Lioness.txt')
+        >>> #Return a network plot for one of the Lioness single sample networks:
+        >>> plot = AnalyzeLioness(lioness_obj)
+        >>> plot.top_network_plot(column= 0, top=100, file='top_100_genes.png')
 
+    Example lioness output:
         TF, Gene and Motif order is identical to the panda output file.
 
-    Authors:
-        Cho-Yi Chen, David Vi, Daniel Morgan
+        - Sample1 Sample2 Sample3 Sample4\n
+        - -------------------------------\n
+        - -0.667452814003	-1.70433776179	-0.158129613892	-0.655795512803\n
+        - -0.843366539284	-0.733709815256	-0.84849895139	-0.915217389738\n
+        - 3.23445386464	2.68888472802	3.35809757371	3.05297381396\n
+        - 2.39500370135	1.84608635425	2.80179804094	2.67540878165\n
+        - -0.117475863987	0.494923925853	0.0518448588965	-0.0584810456421\n
 
-    Reference:
-        Kuijjer, Marieke Lydia, et al. "Estimating sample-specific regulatory networks." Iscience 14 (2019): 226-240.
+        
+
+    References
+    -----------
+
+    .. [1]__ Kuijjer, Marieke Lydia, et al. "Estimating sample-specific regulatory networks." 
+        Iscience 14 (2019): 226-240.
+    
+    Authors: Cho-Yi Chen, David Vi, Daniel Morgan
     """
 
     def __init__(
@@ -72,38 +109,16 @@ class Lioness(Panda):
         save_fmt="npy",
         output="network",
         alpha=0.1,
+        save_single = False,
+        export_filename = None
     ):
-        """
-        Description:
-            Initialize instance of Lioness class and load data.
-
-        Inputs:
-            obj             : PANDA object, generated with keep_expression_matrix=True.
-            obj.motif_matrix: TF DNA motif binding data in tf-by-gene format.
-                              If set to None, Lioness will be performed on gene coexpression network.
-            computing       : 'cpu' uses Central Processing Unit (CPU) to run PANDA
-                              'gpu' use the Graphical Processing Unit (GPU) to run PANDA
-            precision       : 'double' computes the regulatory network in double precision (15 decimal digits).
-                              'single' computes the regulatory network in single precision (7 decimal digits) which is fastaer, requires half the memory but less accurate.
-            start           : Index of first sample to compute the network.
-            end             : Index of last sample to compute the network.
-            save_dir        : Directory to save the networks.
-            save_fmt        : Save format.
-                              '.npy': (Default) Numpy file.
-                              '.txt': Text file.
-                              '.mat': MATLAB file.
-            output          : 'network' returns all networks in a single edge-by-sample matrix (lioness_obj.total_lioness_network). For large sample sizes, this variable requires large RAM memory.
-                              'gene_targeting' returns gene targeting scores for all networks in a single gene-by-sample matrix (lioness_obj.total_lioness_network).
-                              'tf_targeting' returns tf targeting scores for all networks in a single gene-by-sample matrix (lioness_obj.total_lioness_network).
-            alpha            : learning rate, set to 0.1 by default but has to be changed manually to match the learning rate of the PANDA object.
-
-        Output:
-            export_lioness_results : Depeding on the output argument, this can be either all the lioness networks or their gene/tf targeting scores.
+        """ Initialize instance of Lioness class and load data.
         """
         # Load data
         with Timer("Loading input data ..."):
             self.export_panda_results = obj.export_panda_results
             self.expression_matrix = obj.expression_matrix
+            self.expression_samples = obj.expression_samples
             self.motif_matrix = obj.motif_matrix
             self.ppi_matrix = obj.ppi_matrix
             self.correlation_matrix = obj.correlation_matrix
@@ -114,6 +129,7 @@ class Lioness(Panda):
             self.computing = computing
             self.alpha = alpha
             self.n_cores = int(ncores)
+            self.save_single = save_single
             if hasattr(obj, "panda_network"):
                 self.network = obj.panda_network.to_numpy()
             elif hasattr(obj, "puma_network"):
@@ -131,6 +147,7 @@ class Lioness(Panda):
         self.indexes = range(self.n_conditions)[
             start - 1 : end
         ]  # sample indexes to include
+        self.expression_samples = self.expression_samples[start-1:end]
         print("Number of total samples:", self.n_conditions)
         print("Number of computed samples:", len(self.indexes))
         print("Number of parallel cores:", self.n_cores)
@@ -158,7 +175,7 @@ class Lioness(Panda):
                 total_genes = [i for i in gene_names for _ in range(len(tf_names))]
                 indDF = pd.DataFrame([total_tfs, total_genes], index=["tf", "gene"])
                 indDF = indDF.append(
-                    pd.DataFrame(self.total_lioness_network)
+                    pd.DataFrame(self.total_lioness_network, index = self.expression_samples)
                 ).transpose()
             else:  # if equal to None to be specific
                 total_genes1 = gene_names * len(gene_names)
@@ -167,7 +184,7 @@ class Lioness(Panda):
                     [total_genes1, total_genes2], index=["gene1", "gene2"]
                 )
                 indDF = indDF.append(
-                    pd.DataFrame(self.total_lioness_network)
+                    pd.DataFrame(self.total_lioness_network, index = self.expression_samples)
                 ).transpose()
             self.export_lioness_results = indDF
         elif output == "gene_targeting":
@@ -178,15 +195,18 @@ class Lioness(Panda):
             self.export_lioness_results = pd.DataFrame(
                 self.total_lioness_network, columns=tf_names
             ).transpose()
-        self.save_lioness_results()
+        if export_filename:
+            self.export_lioness_table(output_filename = export_filename)
+        else:
+            self.save_lioness_results()
 
     def __lioness_loop(self, i):
-        """
-        Description:
-            Initialize instance of Lioness class and load data.
+        """ Initialize instance of Lioness class and load data.
 
-        Outputs:
-            self.total_lioness_network: An edge-by-sample matrix containing sample-specific networks.
+        Returns
+        --------
+            self.total_lioness_network: array
+                An edge-by-sample matrix containing sample-specific networks.
         """
         # for i in self.indexes:
         print("Running LIONESS for sample %d:" % (i + 1))
@@ -194,7 +214,7 @@ class Lioness(Panda):
         with Timer("Computing coexpression network:"):
             if self.computing == "gpu":
                 import cupy as cp
-
+                
                 correlation_matrix = cp.corrcoef(self.expression_matrix[:, idx])
                 if cp.isnan(correlation_matrix).any():
                     cp.fill_diagonal(correlation_matrix, 1)
@@ -233,22 +253,24 @@ class Lioness(Panda):
         # old
         #lioness_network = self.n_conditions * (self.network - subset_panda_network) + subset_panda_network
 
-        with Timer(
-            "Saving LIONESS network %d to %s using %s format:"
-            % (i + 1, self.save_dir, self.save_fmt)
-        ):
-            path = os.path.join(self.save_dir, "lioness.%d.%s" % (i + 1, self.save_fmt))
-            if self.save_fmt == "txt":
-                np.savetxt(path, lioness_network)
-            elif self.save_fmt == "npy":
-                np.save(path, lioness_network)
-            elif self.save_fmt == "mat":
-                from scipy.io import savemat
+        if self.save_single:
+            with Timer(
+                "Saving LIONESS network %d (%s) to %s using %s format:"
+                % (i + 1, self.expression_samples[i], self.save_dir, self.save_fmt)
+            ):
+                path = os.path.join(self.save_dir, "lioness.%s.%s" % (self.expression_samples[i], self.save_fmt))
+                if self.save_fmt == "txt":
+                    np.savetxt(path, lioness_network)
+                elif self.save_fmt == "npy":
+                    np.save(path, lioness_network)
+                elif self.save_fmt == "mat":
+                    from scipy.io import savemat
 
-                savemat(path, {"PredNet": lioness_network})
-            else:
-                print("Unknown format %s! Use npy format instead." % self.save_fmt)
-                np.save(path, lioness_network)
+                    savemat(path, {"PredNet": lioness_network})
+                else:
+                    print("Unknown format %s! Use npy format instead." % self.save_fmt)
+                    np.save(path, lioness_network)
+
         if self.computing == "gpu" and i == 0:
             self.total_lioness_network = np.fromstring(
                 np.transpose(lioness_network).tostring(), dtype=lioness_network.dtype
@@ -269,12 +291,12 @@ class Lioness(Panda):
     @delayed
     @wrap_non_picklable_objects
     def __par_lioness_loop(self, i, output):
-        """
-        Description:
-            Initialize instance of Lioness class and load data.
+        """ Initialize instance of Lioness class and load data.
 
-        Outputs:
-            self.total_lioness_network: An edge-by-sample matrix containing sample-specific networks.
+        Returns
+        ---------
+            self.total_lioness_network: array
+                An edge-by-sample matrix containing sample-specific networks.
         """
         # for i in self.indexes:
         print("Running LIONESS for sample %d:" % (i + 1))
@@ -321,22 +343,24 @@ class Lioness(Panda):
             (self.n_conditions - 1) * subset_panda_network
         )
 
-        with Timer(
-            "Saving LIONESS network %d to %s using %s format:"
-            % (i + 1, self.save_dir, self.save_fmt)
-        ):
-            path = os.path.join(self.save_dir, "lioness.%d.%s" % (i + 1, self.save_fmt))
-            if self.save_fmt == "txt":
-                np.savetxt(path, lioness_network)
-            elif self.save_fmt == "npy":
-                np.save(path, lioness_network)
-            elif self.save_fmt == "mat":
-                from scipy.io import savemat
+        # if save_single flag is passed, save each single lioness sample
+        if self.save_single:
+            with Timer(
+                "Saving LIONESS network %d (%s) to %s using %s format:"
+                % (i + 1,self.expression_samples[i], self.save_dir, self.save_fmt)
+            ):
+                path = os.path.join(self.save_dir, "lioness.%s.%s" % (self.expression_samples[i], self.save_fmt))
+                if self.save_fmt == "txt":
+                    np.savetxt(path, lioness_network)
+                elif self.save_fmt == "npy":
+                    np.save(path, lioness_network)
+                elif self.save_fmt == "mat":
+                    from scipy.io import savemat
 
-                savemat(path, {"PredNet": lioness_network})
-            else:
-                print("Unknown format %s! File will not be saved." % self.save_fmt)
-                # np.save(path, lioness_network)
+                    savemat(path, {"PredNet": lioness_network})
+                else:
+                    print("Unknown format %s! File will not be saved." % self.save_fmt)
+                    # np.save(path, lioness_network)
         # if i == 0:
         # self.total_lioness_network = np.fromstring(np.transpose(lioness_network).tostring(),dtype=lioness_network.dtype)
         # else:
@@ -349,46 +373,51 @@ class Lioness(Panda):
             self.total_lioness_network = np.sum(lioness_network, axis=1)
         return self.total_lioness_network
 
-    def save_lioness_results(self):
-        """
-        Description:
-            Saves LIONESS network.
-
-        Outputs:
-            file: Path to save the network.
+    def save_lioness_results(self, lioness_output_filename = None):
+        """ Saves LIONESS network.
+            Uses self.save_fmt, self.save_dir to save the data
+            into self.total_lioness_network
         """
         # self.lioness_network.to_csv(file, index=False, header=False, sep="\t")
-        fullpath = os.path.join(self.save_dir, "lioness.%s" % (self.save_fmt))
-        if self.save_fmt == "txt":
+        if lioness_output_filename:
+            fullpath = lioness_output_filename
+        else:
+            fullpath = os.path.join(self.save_dir, "lioness.%s" % (self.save_fmt))
+
+        if fullpath.endswith("txt"):
             np.savetxt(
                 fullpath,
                 np.transpose(self.total_lioness_network),
                 delimiter="\t",
-                header="",
+                header = False,
             )
-        elif self.save_fmt == "npy":
+        elif fullpath.endswith("npy"):
             np.save(fullpath, np.transpose(self.total_lioness_network))
-        elif self.save_fmt == "mat":
+        elif fullpath.endswith("mat"):
             from scipy.io import savemat
 
             savemat(fullpath, np.transpose(self.total_lioness_network))
+        else:
+            print('Trying to save lioness output. Format %s not recognised' %str(fullpath))
         return None
 
     def export_lioness_table(self, output_filename="lioness_table.txt", header=False):
-        """
-        Description:
+        """ 
             Saves LIONESS network with edge names. This saves a dataframe with the corresponding
             header and indexes.
-            So far we
-        Outputs:
-            output_filename: Path to save the network. Specify relative path
-            and format. Choose between .csv, .tsv and .txt. (Defaults to .lioness_table.txt))
+
+        Parameters
+        ------------
+            output_filename: str
+                Path to save the network. Specify relative path
+                and format. Choose between .csv, .tsv and .txt. 
+                (Defaults to .lioness_table.txt))
         """
         df  = self.export_lioness_results
         df = df.sort_values(by=['tf','gene'])
         if output_filename.endswith("txt"):
-            df.to_csv(output_filename, sep=" ", header=False, index = False)
+            df.to_csv(output_filename, sep=" ", index = False)
         elif output_filename.endswith("csv"):
-            df.to_csv(output_filename, sep=",", header=False, index = False)
+            df.to_csv(output_filename, sep=",", index = False)
         elif output_filename.endswith("tsv"):
-            df.to_csv(output_filename, sep="\t", header=False, index = False)
+            df.to_csv(output_filename, sep="\t", index = False)